refactor: change common game components structure
